+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPZNZK9AWYGH9NR049RCM5RG"
+        "01KQ6F63Z37X6C2D8CS7TQPV2G"
     )
     r"""The unique identifier of the data source"""