+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPZNZK990925BW4P2S3BD037"
+        "01KQ6F63Z24DW9GSTZSDNJVCFQ"
     )
     r"""The unique identifier of the data source"""