+++ packages/orq-rc/RELEASES.md
@@ -7329,3 +7329,13 @@ Based on:
 - [python v4.9.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.0 - packages/orq-rc
+
+## 2026-04-27 03:22:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.1 - packages/orq-rc